# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Flutter mobil uygulaması için Tarım Yönetim Sistemi REST API",
        "title": "Tarım Yönetim Sistemi API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Destek",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/dashboard/charts/income-expense": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Aylık gelir-gider grafik verilerini getirir",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Gelir-gider grafik",
                "parameters": [
                    {
                        "enum": [
                            "month",
                            "quarter",
                            "year"
                        ],
                        "type": "string",
                        "description": "Period (month/quarter/year)",
                        "name": "period",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
        "/dashboard/charts/production": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Üretim kategorileri grafik verilerini getirir",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Üretim grafik",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
        "/dashboard/recent-activities": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Son aktiviteleri listeler",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Son aktiviteler",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit (default: 10)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "additionalProperties": true
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
        "/dashboard/summary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Dashboard için özet istatistikleri getirir",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Dashboard özet",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.DashboardSummary"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "models.APIMeta": {
            "type": "object",
            "properties": {
                "requestId": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "models.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/models.APIError"
                },
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/models.APIMeta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "models.AnimalSummary": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "percentage": {
                    "type": "number"
                },
                "trend": {
                    "type": "string"
                }
            }
        },
...
        "models.DashboardSummary": {
            "type": "object",
            "properties": {
                "activeProducts": {
                    "$ref": "#/definitions/models.ProductSummary"
                },
                "monthlyExpense": {
                    "$ref": "#/definitions/models.FinanceSummary"
                },
                "monthlyIncome": {
                    "$ref": "#/definitions/models.FinanceSummary"
                },
                "totalAnimals": {
                    "$ref": "#/definitions/models.AnimalSummary"
                },
                "totalLands": {
                    "$ref": "#/definitions/models.LandSummary"
                }
            }
        },
        "models.FinanceSummary": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "currency": {
                    "type": "string"
                },
                "trend": {
                    "type": "string"
                }
            }
        },
...
        "models.LandSummary": {
            "type": "object",
            "properties": {
                "area": {
                    "type": "number"
                },
                "count": {
                    "type": "integer"
                },
                "productivity": {
                    "type": "number"
                }
            }
        },
...
        "models.ProductSummary": {
            "type": "object",
            "properties": {
                "categories": {
                    "type": "integer"
                },
                "count": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT token ile kimlik doğrulama",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}